feat/refactor: add workouts page, refactor app component
