+++ src/App.jsx
@@ -5,11 +5,8 @@ import User from './pages/user'
 import Workouts from './pages/workouts'
 import Login from './pages/login'
 import workoutService from './services/workouts'
-import WorkoutList from './components/WorkoutList'
-import AddWorkoutForm from './components/forms/AddWorkoutForm'
 import Notification from './components/Notification'
 import {
-  Box,
   Button,
   Typography,
   useMediaQuery,
@@ -57,69 +54,6 @@ function App() {
     }
   }
 
-  function handleAddWorkout(newWorkout) {
-    workoutService
-      .create(newWorkout)
-      .then((returnedWorkout) => {
-        setWorkouts([...workouts, returnedWorkout])
-        setMessage('workout added')
-        setTimeout(() => {
-          setMessage(null)
-        }, 3500)
-      })
-      .then(() => {
-        workoutService.getAll().then((updatedWorkouts) => {
-          setWorkouts(updatedWorkouts)
-        })
-      })
-      .catch((error) => {
-        console.log(error)
-      })
-  }
-
-  function handleUpdateWorkout(id, updatedWorkout) {
-    workoutService
-      .update(id, updatedWorkout)
-      .then((returnedWorkout) => {
-        setWorkouts(workouts.map((w) => (w.id !== id ? w : returnedWorkout)))
-        setMessage('workout updated')
-        setTimeout(() => {
-          setMessage(null)
-        }, 3500)
-      })
-      .catch((error) => console.log(error))
-  }
-
-  function handleDeleteWorkout(id) {
-    workoutService
-      .remove(id)
-      .then(() => {
-        setWorkouts(workouts.filter((w) => w.id !== id))
-        setMessage('workout deleted')
-        setTimeout(() => {
-          setMessage(null)
-        }, 3500)
-      })
-      .catch((error) => console.log(error))
-  }
-
-  const loggedUser = () => {
-    return (
-      <Box sx={{ marginTop: theme.spacing(3) }}>
-        <Button
-          type="button"
-          onClick={handleLogout}
-          variant="contained"
-          color="secondary"
-          sx={{ marginBottom: theme.spacing(3) }}
-        >
-          Logout
-        </Button>
-        <AddWorkoutForm onAddWorkout={handleAddWorkout} />
-      </Box>
-    )
-  }
-
   return (
     <Router>
       <Paper sx={{ padding: theme.spacing(2) }}>
@@ -134,9 +68,20 @@ function App() {
             my account
           </Link>
           {user ? (
-            <Typography variant="body1" gutterBottom>
-              {user.name} logged in
-            </Typography>
+            <>
+              <Typography variant="body1" gutterBottom>
+                {user.name} logged in
+              </Typography>
+              <Button
+                type="button"
+                onClick={handleLogout}
+                variant="contained"
+                color="secondary"
+                sx={{ marginBottom: theme.spacing(3) }}
+              >
+                Logout
+              </Button>
+            </>
           ) : (
             <Link style={{ padding: 5 }} to="/login">
               login
@@ -145,7 +90,17 @@ function App() {
         </nav>
 
         <Routes>
-          <Route path="/workouts" element={<Workouts />} />
+          <Route
+            path="/workouts"
+            element={
+              <Workouts
+                workouts={workouts}
+                user={user}
+                setMessage={setMessage}
+                setWorkouts={setWorkouts}
+              />
+            }
+          />
           <Route path="/account" element={<User />} />
           <Route path="/" element={<Home />} />
           <Route
@@ -161,18 +116,14 @@ function App() {
         <Notification message={message} />
         <ThemeSwitch />
 
-        {user && loggedUser()}
-
         {isLoading ? (
           <Typography variant="body1" gutterBottom>
             Loading data...
           </Typography>
         ) : (
-          <WorkoutList
-            workouts={workouts}
-            onDeleteWorkout={handleDeleteWorkout}
-            onUpdateWorkout={handleUpdateWorkout}
-          />
+          <Typography variant="body1" gutterBottom>
+            Data loaded - info about workouts
+          </Typography>
         )}
       </Paper>
     </Router>
